Updated model for feed
