+++ Listenbrainz/model/Feed/FeedMbidMapping.swift
@@ -12,18 +12,19 @@ import Foundation
 struct FeedMbidMapping: Codable {
     let artistMbids: [String]
     let artists: [FeedArtist]
+    let caaID: Int?
+    let caaReleaseMbid: String?
     let recordingMbid: String
     let recordingName: String?
-    let caaID: Int?
-    let caaReleaseMbid, releaseMbid: String?
+    let releaseMbid: String
 
     enum CodingKeys: String, CodingKey {
         case artistMbids = "artist_mbids"
         case artists
-        case recordingMbid = "recording_mbid"
-        case recordingName = "recording_name"
         case caaID = "caa_id"
         case caaReleaseMbid = "caa_release_mbid"
+        case recordingMbid = "recording_mbid"
+        case recordingName = "recording_name"
         case releaseMbid = "release_mbid"
     }
 }
